+++ src/components/Bird/BirdDetail.scss
@@ -1,7 +1,6 @@
 @import "../../assets/css/variables";
 
 .BirdDetail {
-
   @media (min-width: $screen-sm-min) {
     .Banner {
       position: relative;
