+++ src/components/Sightings/SightingDetail.scss
