+++ src/components/helpers/PrettyBandCombo.js
