+++ src/components/helpers/QualityIndicator.js
