+++ src/components/helpers/QualityIndicator.scss
