+++ src/components/helpers/generateMarker.js
