+++ src/components/presentation/Banner/Banner.scss
@@ -1,4 +1,4 @@
-@import "../../assets/css/variables";
+@import "../../../assets/css/variables";
 
 .Banner {
   background: $gray-lighter;
