+++ src/components/presentation/Footer/Footer.scss
@@ -1,4 +1,4 @@
-@import "../../assets/css/variables";
+@import "../../../assets/css/variables";
 
 footer {
   background-color: $brand-faded;
