+++ src/components/presentation/Navigation/Navigation.scss
@@ -1,4 +1,4 @@
-@import "../../assets/css/variables";
+@import "../../../assets/css/variables";
 
 .navbar.navbar-inverse {
   border: 0;
