+++ src/components/wordpress/Blog.scss
@@ -2,7 +2,7 @@
 
 .BlogPost {
   margin-bottom: $line-height-computed;
-  
+
   h3, h4 {
     font-family: $font-family-base;
   }
