+++ src/views/birds/detail.js
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 
-import BirdProfile from '../../components/Bird/BirdProfile';
+import BirdDetail from '../../components/Bird/BirdDetail';
 import BirdSightingsTable from '../../components/BirdSightings/BirdSightingsTable';
 import BirdSightingsMap from '../../components/BirdSightings/BirdSightingsMap';
 
@@ -8,7 +8,7 @@ class BirdDetailPage extends Component {
   render() {
     return (
       <div className="BirdDetailPage">
-        <BirdProfile slug={this.props.match.params.slug} />
+        <BirdDetail slug={this.props.match.params.slug} />
         <div className="container">
           <h2>Latest Sightings</h2>
           <div className="row">
