+++ src/views/sightings/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import Helmet from 'react-helmet';
 
-import Banner from '../../components/Banner/Banner';
+import Banner from '../../components/presentation/Banner/Banner';
 import SightingsTable from '../../components/Sightings/SightingsTable';
 import SightingsMap from '../../components/Sightings/SightingsMap';
